feat: add REST controllers and refactor native tests for AT mode
- Add REST controllers (Account, Order, Storage, Seata) and DTOs for test-native module
- Replace DataSourceProxyModeAtTests with focused service/controller test classes
- Add test-native app startup step in native build CI workflow
- Add GraalVM native-image agent config generation to Makefile

+++ test-suite/test-native/src/main/java/org/apache/seata/entity/Account.java
@@ -30,13 +30,13 @@ public class Account {
     private String userId;
 
     @Column(columnDefinition = "INT DEFAULT 0")
-    private Integer money;
+    private Long money;
 
     // Constructors
 
     public Account() {}
 
-    public Account(String userId, Integer money) {
+    public Account(String userId, Long money) {
         this.userId = userId;
         this.money = money;
     }
@@ -59,11 +59,11 @@ public class Account {
         this.userId = userId;
     }
 
-    public Integer getMoney() {
+    public Long getMoney() {
         return money;
     }
 
-    public void setMoney(Integer money) {
+    public void setMoney(Long money) {
         this.money = money;
     }
